Updated fitness calculation
