+++ Java/Genetic algorithm visualization/Code12
